% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{barmap_loading}
\alias{barmap_loading}
\title{Barmap of standardized loadings}
\usage{
barmap_loading(
  pca,
  pca_data,
  ndim = 1:5,
  cutoff = 0,
  resample_ci = NULL,
  conf = 0.95,
  plot_list_center = F,
  plot_legend = T,
  text_values = F,
  star_values = F,
  text_size = 2,
  plot_cutoff = T,
  vars = NULL,
  colors = c("steelblue1", "white", "firebrick1"),
  gradient_color = T
)
}
\arguments{
\item{pca}{Object of class \emph{prcomp}, \emph{princals}, or \emph{data.frame}. If object is a \emph{prcomp} or \emph{princals} object, \emph{pca_data} is required, and
the loadings will be extracted. If object is a data.frame object, the dataframe needs to be formatted as: first column named \emph{Variables} and all other columns corresponding to a PC.
One row per variable. The values are the loadings.}

\item{pca_data}{Data passed to the \emph{prcomp} or \emph{princals} function.}

\item{ndim}{Numeric. Number of PCs to plot}

\item{cutoff}{Numeric or numeric vector of length \emph{ndim}.
Value of the loadings threshold (i.e. |loadings| >= cutoff) to plot with stars. Default = 0.5}

\item{resample_ci}{dataframe. Dataframe containing the columns "Variables", "component","original", "mean","ci_low" and "ci_high" containing the center, the lower bound and the upper bound of the confident intervals to plot.
These can be obtained by the pc_stability or permut_pc_test "results" element or computed independently.}

\item{conf}{Numeric. Confidence level used when \emph{load_list} is provided.}

\item{plot_list_center}{Logical. Whether to plot the average loadings obtained from \emph{load_list}.}

\item{plot_legend}{Logical. Whether legend should be plotted.}

\item{text_values}{Logical. Whether to plot the values of the loadings or not. Default=TRUE}

\item{star_values}{Logical. Whether to plot a star in |loadings|>=cutoff. Only relevant if
\emph{text_values}=FALSE. Default=FALSE}

\item{text_size}{Numeric. Size of the text_values.}

\item{plot_cutoff}{Logical. Whether to plot the cutoff lines or not.}

\item{vars}{Character vector. Variables will be ordered as the provided variable names. Non-specified
variables will be excluded from the plot. By default variables are ordered in alphabetically by ggplot.}

\item{colors}{Character vector of length 3. Vector with the character name or
hexadecimal number (e.g. "#FF0000") of three colors, the lower color, the middle color and the higher color
for the gradient used in the plot. Hexadecimal number can be obtained using \emph{rgb} for example.}

\item{gradient_color}{Logical. Whether colors should be plotted as a gradient proportional to the loadings.}
}
\value{
Returns a \emph{ggplot2} object.
}
\description{
Plot a Barmap of the standardized loadings from a PCA solution.
}
\examples{
pca<-prcomp(mtcars, center = TRUE, scale. = TRUE)

}
\author{
Abel Torres Espin
}
