% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract_category_quant_plot}
\alias{extract_category_quant_plot}
\title{Extract the category quantification plot}
\usage{
extract_category_quant_plot(
  pca,
  pca_data,
  var,
  plot_dim = c(1, 2),
  nudge_y,
  nudge_x
)
}
\arguments{
\item{pca}{Object of class \emph{prcomp}, \emph{princals}, or \emph{data.frame}. If object is a \emph{prcomp} or \emph{princals} object, \emph{pca_data} is required, and
the loadings will be extracted. If object is a data.frame object, the dataframe needs to be formatted as: first column named \emph{Variables} and all other columns corresponding to a PC.
One row per variable. The values are the loadings.}

\item{pca_data}{Data passed to the \emph{prcomp} or \emph{princals} function.}

\item{var}{Character. Character name of the variables to plot.}

\item{plot_dim}{Numeric vector of length 2. Dimensions (aka principal components) to be plotted.}

\item{nudge_y}{Numeric. Controls y the displacement of the label position for the name of each level.}

\item{nudge_x}{Numeric. Controls y the displacement of the label position for the name of each level.}
}
\value{
Returns a list \emph{ggplot2} object, one per each specified variable
}
\description{
Plot of the projection of the category quantification into the loading vector for
non-linear pca variables of an object of the class "princals".
}
\examples{
data(mtcars)
pca_mtcars<-Gifi::princals(mtcars)

extract_category_quant_plot(pca = pca_mtcars, pca_data = mtcars, var="cyl"))

}
\references{
\enumerate{
 \item Linting, M., Meulman, J. J., Groenen, P. J. F., & van der Koojj, A. J. (2007). Nonlinear principal components analysis: Introduction and application. Psychological Methods, 12(3), 336–358. https://doi.org/10.1037/1082-989X.12.3.336
 \item Linting, M., & Kooij, A. van der. (2012). Nonlinear Principal Components Analysis With CATPCA: A Tutorial. Journal of Personality Assessment, 94(1), 12–25. https://doi.org/10.1080/00223891.2011.627965
}
}
\author{
Abel Torres Espin
}
