% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract_loadings}
\alias{extract_loadings}
\title{extract loadings}
\usage{
extract_loadings(pca, pca_data)
}
\arguments{
\item{pca}{Object of class \emph{prcomp}, \emph{princals} or \emph{data.frame}. If the object is a \emph{prcomp} or \emph{princals} object, \emph{pca_data} is required, and
the loadings will be extracted. If the object is a data.frame object, the dataframe needs to be formatted as: first column named \emph{Variables} and all other columns corresponding to a PC.
One row per variable. The values are the loadings.}

\item{pca_data}{Data passed to the \emph{prcomp} or \emph{princals} function.}
}
\value{
A data.frame with the standardized loadings in the form of variables as rows and components as columns.
}
\description{
This is a wrapper function for \emph{stand_loadings()}
with added functionalities such as error breakers that is used by
most functions in the package.
}
\author{
Abel Torres Espin
}
