% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract_rmse}
\alias{extract_rmse}
\title{Extracts root mean square error (RMSE)}
\usage{
extract_rmse(vector1, vector2)
}
\arguments{
\item{vector1}{First numerical vector of loadings for the calculation.}

\item{vector2}{Second numerical vector of loadings for the calculation.}
}
\value{
Returns the root mean square (RMS) between vector1 and vector2.
}
\description{
Given two vectors of loadings, computes the root mean square between them.
}
\details{
RMSE has been also used as a metric
  for factor matching (see ref). It is calculated as:
  \deqn{RMSE_{x,y} = sqrt( sum((x_{i}-y_{i})^2) / n)} Where \emph{n} is the number of variables
  in both components \emph{x} and \emph{y}. A RMSE of 0 corresponds to a perfect match.
  The smaller the RMSE is, the more equivalent two components are.
}
\references{
Guadagnoli E, Velicer W. A Comparison of Pattern Matching Indices. Multivar Behav Res. 1991 Apr;26(2):323–43
}
\author{
Abel Torres Espin
}
