% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract_syndromic_plot}
\alias{extract_syndromic_plot}
\title{Extract the syndromic plot}
\usage{
extract_syndromic_plot(
  load_df,
  pc,
  cutoff = 0.5,
  VAF,
  arrow_size_multi = 10,
  repel = F,
  plot_legend = T,
  text_size = 6,
  var_order = "abs decreasing",
  colors = c("steelblue1", "white", "firebrick1")
)
}
\arguments{
\item{load_df}{data.frame with the loadings. Format: one column named \emph{Variable} and a column
for each PC. It can be extracted from a \emph{prcomp} or \emph{princals} object using \emph{stand_loadings()}.}

\item{pc}{String. Name of the PC in \emph{load_df} to plot.}

\item{cutoff}{Numeric. Value of the loadings threshold (i.e. |loadings| > cutoff) to plot. Default = 0.5}

\item{VAF}{String. Text for the center of the syndromic plot. The text generally corresponds to the variance accounted
for (VAF) of the PC. If pca is the results of \emph{prcomp} or \emph{princals}, VAF is internally calculated.}

\item{arrow_size_multi}{Numeric. Controls the size of the arrows proportional to the loading. Default=10}

\item{repel}{Logical. Whether to repel the text for preventing text overlap. Default=FALSE}

\item{plot_legend}{Logical. Whether to plot the legend or not. Default=TRUE}

\item{text_size}{Numeric. Controls for the size of the text. Default=9}

\item{var_order}{Character, character vector or numeric vector. Specify the order of the variables in the plot by the loading values, starting at 12 o’clock and moving counterclockwise.
Possible values: 'abs decreasing': plot by decreasing absolute value;
'abs increasing': plot by increasing absolute value; 'decreasing'; or 'increasing’. A vector can be specified with a custom order of the variables
to plot. If a numeric vector is provided, the variables will be ordered as specified by the numbers. If a character
vector is provided, the variables will be ordered in the same order as the variable names provided. In case of vector, non-specified
variables will be excluded from the plot.}

\item{colors}{Character vector of length 3. Vector with the character name or
hexadecimal number (e.g. "#FF0000") of three colors, the lower color, the middle color and the higher color
for the gradient used in the plot. Hexadecimal number can be obtained using \emph{rgb} for example.}
}
\value{
A \emph{ggplot2} object of the syndromic plot.
}
\description{
Extract the syndromic plot of a given list of loadings and specified pc.
}
\author{
Abel Torres Espin
}
