% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{heatmap_loading}
\alias{heatmap_loading}
\title{Heatmap of standardized loadings}
\usage{
heatmap_loading(
  pca,
  pca_data,
  ndim = 10,
  cutoff = 0.5,
  arbitrary_var = NULL,
  plot_title = "Standardized loadings",
  legend_title = "s. loading",
  text_values = T,
  star_values = F,
  text_size = 2,
  vars = NULL
)
}
\arguments{
\item{pca}{Object of class \emph{prcomp}, \emph{princals}, or \emph{data.frame}. If object is a \emph{prcomp} or \emph{princals} object, \emph{pca_data} is required, and
the loadings will be extracted. If object is a data.frame object, the dataframe needs to be formatted as: first column named \emph{Variables} and all other columns corresponding to a PC.
One row per variable. The values are the loadings.}

\item{pca_data}{Data passed to the \emph{prcomp} or \emph{princals} function.}

\item{ndim}{Numeric. Number of PCs to plot}

\item{cutoff}{Numeric or numeric vector of length \emph{ndim}.
Value of the loadings threshold (i.e. |loadings| >= cutoff) to plot with stars. Default = 0.5}

\item{arbitrary_var}{Character or character vector with the names of the variables where the loadings should be plot as absolute values.
This is the case for categorical variables in categorical PCA where variables do not have direction.}

\item{plot_title}{String. Title of the plot. 'Standardized loadings' by default.}

\item{legend_title}{String. Title of the legend. 's. loading' by default.}

\item{text_values}{Boolean. Whether to plot the values of the loadings or not. Default=TRUE}

\item{star_values}{Boolean. Whether to plot a star in |loadings|>=cutoff. Only relevant if
\emph{text_values}=FALSE. Default=FALSE}

\item{text_size}{Numeric. Size of the text_values.}

\item{vars}{Character vector. Variables will be ordered as the provided variable names. Non-specified
variables will be excluded from the plot. By default variables are ordered in alphabetically by ggplot.}
}
\value{
Returns a \emph{ggplot2} object.
}
\description{
Plot a heatmap of the standardized loadings from a PCA solution.
}
\author{
Abel Torres Espin
}
