% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{new_syndromics}
\alias{new_syndromics}
\title{syndromic class constructor}
\usage{
new_syndromics(x)
}
\arguments{
\item{x}{Object of class list to structure as "syndromics" class}
}
\description{
constructor for the "syndromics" class
}
\author{
Abel Torres Espin
}
