% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{pc_stability}
\alias{pc_stability}
\title{PC stability by nonparametric bootstrapping}
\usage{
pc_stability(
  pca,
  pca_data,
  ndim = 3,
  B = 1000,
  sim = "ordinary",
  communalities = T,
  test_similarity = T,
  similarity_metric = "all",
  s_cut_off = 0.1,
  ci_type = "bca",
  conf = 0.95,
  barmap_plot = T,
  ...
)
}
\arguments{
\item{pca}{Object of class \emph{prcomp}, \emph{princals}, or \emph{data.frame}. If object is a \emph{prcomp} or \emph{princals} object, \emph{pca_data} is required, and
the loadings will be extracted. If object is a data.frame object, the dataframe needs to be formatted as: first column named \emph{Variables} and all other columns corresponding to a PC.
One row per variable. The values are the loadings.}

\item{pca_data}{Data passed to the \emph{prcomp} or \emph{princals} function.}

\item{ndim}{Numeric. Number of PCs (1 to \emph{ndim}) to run the analysis on. Default = 3.}

\item{B}{Numeric. Number of bootstrapped samples passed to \emph{boot}.}

\item{sim}{Character. Determines the bootstrapping method for \emph{boot}.
From \emph{boot}: A character string indicating the type of simulation required.
Possible values are "ordinary" (the default), "parametric", "balanced",
"permutation", or "antithetic". Default="ordinary".}

\item{communalities}{Boolean. Whether to compute and return communalities.}

\item{test_similarity}{Boolean. Whether to compute component similarity by the
specified \emph{similarity_metric} and the nonparametric percentile confident interval.
See \emph{?component_similarity} for more information. Default=TRUE}

\item{similarity_metric}{character or character vector. Specify the similarity metric
to use when \emph{test_similarity}=TRUE. Possible values are "cc_index" (congruence coefficient),
"r_correlation" (Pearson's r), "rmse" (root mean squared error), "s_index' (Cattell's s metric), or "all".
See \emph{?component_similarity} for more details on the available metrics. Default="all".}

\item{s_cut_off}{Numeric. This is the loading cut off used to determine if a
variable is silent or not in Cattell's terms. See \emph{?extract_s} for more
information. Default=0.1.}

\item{ci_type}{Character. Type of confidence interval to compute. This argument is passed to
the \emph{boot.ci} function from the \emph{boot} package. See \emph{?boot.ci} for options. Given
that the BCA method has demonstrated good performance for bootstrapping PCAs,
we have set 'bca' as the default. See ref for more details.}

\item{conf}{Numeric. Level of confidence region for the confidence interval. E.g. 0.95 generates 95CI. Default=0.95}

\item{barmap_plot}{Boolean. Whether to generate a barmap plot of the bootstrapped loadings or not.
See \emph{?barmap_loadings} for details. Default=TRUE}

\item{...}{Other arguments passed to the \emph{barmap_loadings} or \emph{barmap_commun} function when \emph{barmap_plot}=TRUE}
}
\value{
Returns a list object.
\describe{
  \item{\strong{boot_sample}}{A list of length B containing the resampled loadings.}
  \item{\strong{results}}{
    \describe{
       \item{\strong{boot_mean}}{A numeric matrix containing the mean loadings of the bootstrapped sample.}
       \item{\strong{ci_low and ci_high}}{A numeric matrix with the lower and upper CI respectively.}
    }}
  \item{\strong{pc_similarity}}{A list of results when \emph{test_similarity}=TRUE.}
  \describe{
   \item{\strong{similarity_mean}}{A numerix matrix with the mean of the chosen similarity metric.}
   \item{\strong{similarity_ci_low and similarity_ci_high}}{A numeric matrix with the lower and upper CI respectively.}
   }
 \item{\strong{boot_barmap_loadings}}{A ggplot2 object with a barmap plot of loadings of the bootstrapped solution.}
 \item{\strong{boot_barmap_loadings}}{A ggplot2 object with a barmap plot of communalities of the bootstrapped solution.}
}
}
\description{
Extract nonparametric estimate of the standardized loadings and the confident
region by means of bootstrapping. This function uses the \emph{boot} function from
the \strong{boot} package.
}
\details{
The number of bootstrap samples is set to 1000 by default, as it has been shown to be a robust number in most conditions of
data complexity and sample size. The user must be careful on setting such number too low which would reduce the performance
of the approximation. However, values that are too high might unnecessarily increase computing time with little gain (REFs).
}
\references{
Efron B. Better Bootstrap Confidence Intervals. J Am Stat Assoc. 1987 Mar 1;82(397):171–85.
}
\author{
Abel Torres Espin
}
