% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{permut_pca}
\alias{permut_pca}
\title{Permutation PCA}
\usage{
permut_pca(
  pca,
  pca_data,
  P = 1000,
  ndim,
  output = "VAF",
  pb = NULL,
  perm.method = "permD"
)
}
\arguments{
\item{pca}{Object of class \emph{prcomp}, \emph{princals} or \emph{data.frame}. If the object is a \emph{prcomp} or \emph{princals} object, \emph{pca_data} is required, and
the loadings will be extracted. If the object is a data.frame object, the data.frame needs to be formatted as: first column named \emph{Variables} and all other columns corresponding to a PC.
One row per variable. The values are the loadings.}

\item{pca_data}{Data passed to the \emph{prcomp} or \emph{princals} function.}

\item{P}{Numeric. Number of permutations to run calling the \emph{permuted_pca} function. Default=1000}

\item{ndim}{Numeric. Number of PCs to save (1 to ndim).}

\item{output}{Character. Determines the output to compute. Possible values are
Variance accounted for ("VAF") or the standardized loadings ("s.loadings"). Default="VAF"}

\item{pb}{Object of class "Progress_bar" "R6" generated by \emph{progress::progress_bar$new()}. Not required.}

\item{perm.method}{Character determining the permutation method to use as in Linting et al., 2011.
"permD" (Buja & Eyuboglu, 1992; Linting et al., 2011) where variables as permuted independently and
concomitantly as opposite of the "permV" permutation strategy (Linting et al., 2011) where
variables are permuted one at the time. If statistic is set to "VAF", \emph{perm.method} "permD" will be
used. Default="permV"}
}
\value{
Returns a list of objects in which each element is a permutated PCA solution of the
specified output.
}
\description{
Generates a list of permutation samples from a PCA that can be used for
nonparametric permutation test (see \emph{permut_pc_test}).
}
\details{
This is a helper function internally called by
\emph{permut_pca_test()} to produce P permutations of the given output of
the \emph{prcomp()} or the \emph{princals()} functions. It returns a
list of the results of permuting the data, conducting a PCA and
extracting either the VAF or the standardized loadings for each P.
}
\references{
\enumerate{
 \item Buja A, Eyuboglu N. Remarks on Parallel Analysis. Multivar Behav Res. 1992 Oct 1;27(4):509–40
 \item Linting M, van Os BJ, Meulman JJ. Statistical Significance of the Contribution of Variables to the PCA solution: An Alternative Permutation Strategy. Psychometrika. 2011 Jul 1;76(3):440–60
}
}
\author{
Abel Torres Espin
}
