% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{permut_pca_D}
\alias{permut_pca_D}
\title{Generic for permutating PCA using permD method}
\usage{
permut_pca_D(pca, ...)
}
\arguments{
\item{pca}{Object of class \emph{prcomp} or \emph{princals}.}

\item{...}{Further arguments pass to \emph{?permut_pca_D.prcomp} or \emph{?permut_pca_D.princals}}
}
\description{
Generic S3 function for permut_pca methods using the permD permutation
method. Current implemented methods are for objects of class "prcomp" returned by the \emph{prcomp()} function, and
objects of class "princals" returned by the \emph{Gifi::princals()} function.
See \emph{?permut_pca_D.prcomp} or \emph{?permut_pca_D.princals} for more details.
}
\details{
Nonparametric permutation for hypothesis testing of the VAF of component, the loadings or communalities
have been studied (see refs). The hypothesis test is defined as:
H(null): PC metric (either VAF or loading) is indistinguishable from a random generation
H(alternative): PC metric (either VAF or loading) is different from random
The null distribution is generated by permuting the values of each variable several times (P)
and re-running the PCA on each permuted sample. Confidence intervals of the permuted distribution (null distribution) are
calculated using the percentile method. The p values are calculated as p = ((q+1))⁄((P+1)), where q is the number of times
the chosen metric is higher in the permuted distribution than in the original PCA solution and P is the number of permutations.
The user should note that the lowest p value that can be calculated is dependent on P. As an example, if P is set to a value of 10 (a relatively low value),
the smallest p value that can be detected is 0.09, considering q=0. Accordingly, P should be set high enough to reach the desired floor p value.
By default, we have set the number of permutations to 1000 (smallest p value approximately equal to 0.001 as a result) as this has been shown to be
high enough for approximating the null distribution in most cases.

Permutation test of the loadings as in (Buja & Eyuboglu, 1992; Peres-Neto et al., 2003) that can serve to determine
the loading threshold, where the variables are permuted simultaneously and concomitantly. Linting et al., designed and tested an strategy
where only one variable is permuted at the time, showing great results in determining the contribution of variables using communalities (Linting et al., 2011).
This method has resulted in better determination of the significant contribution of variables on the PCA solution with higher statistical power and proper
type I error, and therefore has been incorporated in the package as the suggested method for loadings and communalities.
Following Linting et al., terminology, user can specify the permutation strategy for the loadings as one variable at the time (\emph{permV}, as in Linting et al., 2011)
or as all the variable together (\emph{permD}, as in Buja & Eyuboglu, 1992; Peres-Neto et al., 2003).
}
\references{
\enumerate{
 \item Buja A, Eyuboglu N. Remarks on Parallel Analysis. Multivar Behav Res. 1992 Oct 1;27(4):509–40
 \item Linting M, van Os BJ, Meulman JJ. Statistical Significance of the Contribution of Variables to the PCA solution: An Alternative Permutation Strategy. Psychometrika. 2011 Jul 1;76(3):440–60
}
}
\author{
Abel Torres Espin
}
