% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot.syndromics}
\alias{plot.syndromics}
\title{Method to plot results from pc_stability or permut_pca_test}
\usage{
\method{plot}{syndromics}(
  x,
  plot_type = "barmap",
  ndim = NULL,
  plot_resample = F,
  communalities = F,
  ...
)
}
\description{
Method to plot results from pc_stability or permut_pca_test
}
