% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{stand_loadings}
\alias{stand_loadings}
\title{standardized loadings}
\usage{
stand_loadings(pca, pca_data)
}
\arguments{
\item{pca}{Object of class \emph{prcomp}, \emph{princals}, or \emph{data.frame}. If object is a \emph{prcomp} or \emph{princals} object, \emph{pca_data} is required, and
the loadings will be extracted. If object is a data.frame object, the dataframe needs to be formatted as: first column named \emph{Variables} and all other columns corresponding to a PC.
One row per variable. The values are the loadings.}

\item{pca_data}{Data passed to the \emph{prcomp} or \emph{princals} function.}
}
\value{
A data.frame with the standardized loadings in the form of variables as rows and components as columns.
}
\description{
Extract the standardized loadings from a \emph{prcomp} object by correlating the PC scores
and the original data.
}
\details{
The standardized loadings are calculated as the eigenvectors times the square roots of the respective eigenvalues and divided by
the variable standard deviation (which is 1 in case of standardized PCA (from correlation matrix)). These are equivalent to
the Pearson's correlation between the pca scores and the original dataset. This is the correlations of the PC
with the variables and the same as the correlation of vector coefficients suggested by Jackson and Hearne in 1987.

This function extracts the standardized loadings from the output of the \emph{prcomp()} or the \emph{princals()} functions. In the case of the
\emph{prcomp()} solution, the standardized loadings are calculated as: s.loadings = eigenvectors x sqrt(eigenvalues) if the PCA was performed on the
standardized (scaled to unit variance) data or s.loadings=(eigenvector x sqrt(eigenvalues))⁄S where S is the vector of the variables standard deviation.
In the case of \emph{princals()}, standardized loadings are returned directly in its output and therefore \emph{stand_loadings()} returns those.
}
\references{
Jackson JE, Hearne FT. Relationships Among Coefficients of Vectors Used In Principal Components. Technometrics. 1973 Aug 1;15(3):601–10.
}
\author{
Abel Torres Espin
}
