% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{barmap_loading}
\alias{barmap_loading}
\title{Barmap of standardized loadings}
\usage{
barmap_loading(
  pca,
  pca_data,
  ndim = 10,
  cutoff = 0.5,
  arbitrary_var = NULL,
  load_list = NULL,
  conf = 0.95,
  plot_list_original = F,
  plot_list_center = F,
  plot_title = "Standardized loadings",
  legend_title = "s. loading",
  text_values = F,
  star_values = T,
  text_size = 2,
  plot_cutoff = T,
  vars = NULL,
  colors = c("steelblue1", "firebrick1")
)
}
\arguments{
\item{pca}{Object of class \emph{prcomp}, \emph{princals}, or \emph{data.frame}. If object is a \emph{prcomp} or \emph{princals} object, \emph{pca_data} is required, and
the loadings will be extracted. If object is a data.frame object, the dataframe needs to be formatted as: first column named \emph{Variables} and all other columns corresponding to a PC.
One row per variable. The values are the loadings.}

\item{pca_data}{Data passed to the \emph{prcomp} or \emph{princals} function.}

\item{ndim}{Numeric. Number of PCs to plot}

\item{cutoff}{Numeric or numeric vector of length \emph{ndim}.
Value of the loadings threshold (i.e. |loadings| >= cutoff) to plot with stars. Default = 0.5}

\item{arbitrary_var}{Character or character vector with the names of the variables where the loadings should be plot as absolute values.
This is the case for categorical variables in categorical PCA where variables do not have direction.}

\item{load_list}{List. List of loading matrices used to plot percentile confidence intervals around the average. If NULL, no error bars are ploted. Default=NULL}

\item{conf}{Numeric. Confidence level used when \emph{load_list} is provided.}

\item{plot_list_original}{Boolean. Whether to plot the loadings obtained from the values passed to \emph{pca} and \emph{pca_data} when \emph{load_list} is not NULL.}

\item{plot_list_center}{Boolean. Whether to plot the average loadings obtained from \emph{load_list}.}

\item{plot_title}{String. Title of the plot. 'Standardized loadings' by default.}

\item{legend_title}{String. Title of the legend. 's. loading' by default.}

\item{text_values}{Boolean. Whether to plot the values of the loadings or not. Default=TRUE}

\item{star_values}{Boolean. Whether to plot a star in |loadings|>=cutoff. Only relevant if
\emph{text_values}=FALSE. Default=FALSE}

\item{text_size}{Numeric. Size of the text_values.}

\item{plot_cutoff}{Boolean. Whether to plot the cutoff lines or not.}

\item{vars}{Character vector. Variables will be ordered as the provided variable names. Non-specified
variables will be excluded from the plot. By default variables are ordered in alphabetically by ggplot.}

\item{colors}{Character vector of length 2. Vector with the character name or
hexadecimal number (e.g. "#FF0000") of two colors, the lower color and the higher color
for the gradient used in the plot. Hexadecimal number can be obtained using \emph{rgb} for example.}
}
\value{
Returns a \emph{ggplot2} object.
}
\description{
Plot a Barmap of the standardized loadings from a PCA solution.
}
\author{
Abel Torres Espin
}
