% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{pc_stability}
\alias{pc_stability}
\title{PC stability by nonparametric bootstrapping}
\usage{
pc_stability(
  pca,
  pca_data,
  ndim = 3,
  B = 1000,
  sim = "ordinary",
  communalities = TRUE,
  similarity_metric = "all",
  s_cut_off = 0.1,
  ci_type = "bca",
  conf = 0.95,
  inParallel = F,
  n_cores = 2,
  inParallel_extra = NULL,
  ...
)
}
\arguments{
\item{pca}{Object of class \emph{prcomp}, \emph{princals}.}

\item{pca_data}{Data passed to the \emph{prcomp} or \emph{princals} function.}

\item{ndim}{Numeric. Number of PCs (1 to \emph{ndim}) to run the analysis on. Default = 3.}

\item{B}{Numeric. Number of bootstrapped samples passed to \emph{boot}.}

\item{sim}{Character. Determines the bootstrapping method for \emph{boot}.
From \emph{boot}: A character string indicating the type of simulation required.
Possible values are "ordinary" (the default), "parametric", "balanced",
"permutation", or "antithetic". Default="ordinary".}

\item{communalities}{Logical. Whether to compute and return communalities.}

\item{similarity_metric}{character or character vector. Specify the similarity metric
to use. Set "none" to not return similarity metrics. Possible values are "cc_index" (congruence coefficient),
"r_correlation" (Pearson's r), "rmse" (root mean squared error), "s_index' (Cattell's s metric), or "all".
See \emph{?component_similarity} for more details on the available metrics. Default="all".}

\item{s_cut_off}{Numeric. This is the loading cut off used to determine if a
variable is silent or not in Cattell's terms. See \emph{?extract_s} for more
information. Default=0.1.}

\item{ci_type}{Character. Type of confidence interval to compute. This argument is passed to
the \emph{boot.ci} function from the \emph{boot} package. See \emph{?boot.ci} for options. Given
that the BCA method has demonstrated good performance for bootstrapping PCAs,
we have set 'bca' as the default. See ref for more details.}

\item{conf}{Numeric. Level of confidence region for the confidence interval. E.g. 0.95 generates 95CI. Default=0.95}

\item{inParallel}{Logical. Whether to run the function in parallel using \emph{pbapply::pblapply}. In window machines,
parallelization is done through \emph{parallel::parLapply}. In Unix machines, parallelization is done through
\emph{parallel::mclapply}. See \emph{?pblapply} for details. Default = F.}

\item{n_cores}{Numeric. Number of cores to use. Available cores can be obtained by \emph{parallel::detectCores()}.
It is recommended to use less cores than available. Default = 2.}

\item{inParallel_extra}{Character or vector Character with the string name of additional objects passed to the clusters when
performing bootstrapping. This might be needed when the function to run the original PCA has been called with external objects.
E.g., princals(..., ndim = ncol(dataFrame)), where dataFrame is a data frame object. In that case \emph{inParallel_extra} = "dataFrame".}

\item{...}{Other arguments passed to the \emph{boot} function.}
}
\value{
Returns a list object of class "syndromics".
\describe{
  \item{\strong{method}}{Specify the method used to obtain the syndromics list object}
  \item{\strong{pca}}{Contains the object passed to the pca argument}
  \item{\strong{pca_data}}{Contains the object passed to the pca_data argument}
  \item{\strong{ndim}}{Value specified in the ndim argument}
  \item{\strong{ci_method}}{Method used to compute CIs}
  \item{\strong{conf}}{Confidence level used to compute CIs}
  \item{\strong{results}}{Object containing the results of the analysis}
  \item{\strong{boot_sample}}{A list of length B containing the resampled loadings.}
  \item{\strong{pc_similarity}}{A list of results when \emph{similarity_metric} is not "none".}
  \describe{
   \item{\strong{similarity_mean}}{A numerix matrix with the mean of the chosen similarity metric.}
   \item{\strong{similarity_ci_low and similarity_ci_high}}{A numeric matrix with the lower and upper CI respectively.}
   }
  \item{\strong{B}}{Number of resamples computed.}
  \item{\strong{communalities}}{If communalities= TRUE, the results are returned here.}
}
}
\description{
Extract nonparametric estimate of the standardized loadings and the confident
region by means of bootstrapping. This function uses the \emph{boot} function from
the \strong{boot} package.
}
\details{
The number of bootstrap samples is set to 1000 by default, as it has been shown to be a robust number in most conditions of
data complexity and sample size. The user must be careful on setting such number too low which would reduce the performance
of the approximation. However, values that are too high might unnecessarily increase computing time with little gain (REFs).
}
\examples{
data(mtcars)
pca_mtcars<-prcomp(mtcars, center = TRUE, scale. = TRUE)

pca_mtcars_stab<-pc_stability(pca = pca_mtcars, pca_data = mtcars, ndim = 3, B = 500)
plot(pca_mtcars_stab, plot_resample= TRUE)

}
\references{
Efron B. Better Bootstrap Confidence Intervals. J Am Stat Assoc. 1987 Mar 1;82(397):171–85.
}
\author{
Abel Torres Espin
}
