% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{permut_pc_test}
\alias{permut_pc_test}
\title{permutation test of PCA}
\usage{
permut_pc_test(
  pca,
  pca_data,
  P = 1000,
  ndim = 3,
  statistic = "VAF",
  conf = 0.95,
  adj.method = "BH",
  perm.method = "permV",
  plot = T,
  colors = c("steelblue", "orange"),
  ...
)
}
\arguments{
\item{pca}{Object of class \emph{prcomp}, \emph{princals}, or \emph{data.frame}. If object is a \emph{prcomp} or \emph{princals} object, \emph{pca_data} is required, and
the loadings will be extracted. If object is a data.frame object, the dataframe needs to be formatted as: first column named \emph{Variables} and all other columns corresponding to a PC.
One row per variable. The values are the loadings.}

\item{pca_data}{Data passed to the \emph{prcomp} or \emph{princals} function.}

\item{P}{Numeric. Number of permutations to run calling the \emph{permuted_pca} function. Default=1000}

\item{ndim}{Numeric. Number of PCs (1 to \emph{ndim}) to run the analysis on. D}

\item{statistic}{Character. Determines the statistic to compute. Possible values are
Variance accounted for ("VAF") or the standardized loadings ("s.loadings"). Default="VAF"}

\item{conf}{Numeric. Level of confidence region for the confidence interval. E.g. 0.95 generates 95CI. Default=0.95}

\item{adj.method}{Character passed to the \emph{stats::p.adjust()} to adjust the p value for multiple comparisons.
See \emph{?p.adjust.methods}. Default="BH"}

\item{perm.method}{Character determining the permutation method to use as in Linting et al., 2011.
"permD" (Buja & Eyuboglu, 1992; Linting et al., 2011) where variables as permuted independently and
concomitantly (Fig. 2A) as opposite of the "permV" permutation strategy (Linting et al., 2011) where
variables are permuted one at the time. If statistic is set to "VAF", \emph{perm.method} "permD" will be
used. Default="permV".}

\item{plot}{Boolean. Whether to rerun a ggplot2 object containing a plot of "VAF", "loadings" or "communalities" as specified in \emph{statistic}.}

\item{colors}{Character vector of length 2. Vector with the character name or
hexadecimal number (e.g. "#FF0000") of two colors, the first color is used for the original VAF, the second color
is used for the permuted VAF and error bars for the gradient used in the plot.
Hexadecimal number can be obtained using \emph{rgb} for example.}

\item{...}{Other arguments passed to the \emph{barmap_loadings} or \emph{barmap_commun} function when \emph{barmap_plot}=TRUE}
}
\value{
List containing the following objects
\describe{
  \item{perm.methods}{Character for the used permutation method}
  \item{per_list}{List of P loading matrices for the P permuted samples}
  \item{statistic}{Character for the used statistic}
  \item{adj.method}{Character for the used method for adjusting p values}
  \item{per_plot}{ggplot2 object with the plot of VAF, loadings or communalities depending on the specified statistic}
  \item{results}{List of data.frame(s) with the results of the permutation test}
}
}
\description{
Compute a nonparametric permutation test for a PCA solution. Two options are
possible: hypothesis testing of the total variance accounted for (VAF) for each component as in XXX ref, and
hypothesis testing of the standardized loadings as in XXX ref.
}
\details{
Nonparametric permutation for hypothesis testing of the VAF of component, the loadings or communalities
have been studied (see refs). The hypothesis test is defined as:
H(null): PC metric (either VAF or loading) is indistinguishable from a random generation
H(alternative): PC metric (either VAF or loading) is different from random
The null distribution is generated by permuting the values of each variable several times (P)
and re-running the PCA on each permuted sample. Confidence intervals of the permuted distribution (null distribution) are
calculated using the percentile method. The p values are calculated as p = ((q+1))⁄((P+1)), where q is the number of times
the chosen metric is higher in the permuted distribution than in the original PCA solution and P is the number of permutations.
The user should note that the lowest p value that can be calculated is dependent on P. As an example, if P is set to a value of 10 (a relatively low value),
the smallest p value that can be detected is 0.09, considering q=0. Accordingly, P should be set high enough to reach the desired floor p value.
By default, we have set the number of permutations to 1000 (smallest p value approximately equal to 0.001 as a result) as this has been shown to be
high enough for approximating the null distribution in most cases.

Permutation test of the loadings as in (Buja & Eyuboglu, 1992; Peres-Neto et al., 2003) that can serve to determine
the loading threshold, where the variables are permuted simultaneously and concomitantly. Linting et al., designed and tested an strategy
where only one variable is permuted at the time, showing great results in determining the contribution of variables using communalities (Linting et al., 2011).
This method has resulted in better determination of the significant contribution of variables on the PCA solution with higher statistical power and proper
type I error, and therefore has been incorporated in the package as the suggested method for loadings and communalities.
Following Linting et al., terminology, user can specify the permutation strategy for the loadings as one variable at the time (\emph{permV}, as in Linting et al., 2011)
or as all the variable together (\emph{permD}, as in Buja & Eyuboglu, 1992; Peres-Neto et al., 2003).
}
\references{
\enumerate{
 \item Buja A, Eyuboglu N. Remarks on Parallel Analysis. Multivar Behav Res. 1992 Oct 1;27(4):509–40
 \item Linting M, van Os BJ, Meulman JJ. Statistical Significance of the Contribution of Variables to the PCA solution: An Alternative Permutation Strategy. Psychometrika. 2011 Jul 1;76(3):440–60
}
}
\author{
Abel Torres Espin
}
